Regrid unit tests house-keeping.
